Move remaining helpers out of test/support and into spec/support - some helpers have already been moved over in previous prs. This moves those that remained. - update require statements with new location of helpers.
