+++ features/support/helpers.rb
@@ -1,6 +1,6 @@
 require "gds_api/test_helpers/content_store"
 require "gds_api/test_helpers/search"
-require_relative "../../test/support/search_api_helpers"
+require_relative "../../spec/support/search_api_helpers"
 
 module TestHelpers
   include GdsApi::TestHelpers::Search
