+++ spec/spec_helper.rb
@@ -9,11 +9,6 @@ require "gds_api/test_helpers/search"
 require "gds_api/test_helpers/content_store"
 require "webmock/rspec"
 
-# The helpers in the test/support directory are used by minitest specs that are not yet
-# converted to RSpec. We should require those helpers rather than duplicate them in
-# spec/support so that the files don't diverge. Once all the specs using a helper are
-# converted to rspec, we can move the helper file over to spec/support.
-Dir[Rails.root.join("test/support/**/*.rb")].sort.each { |f| require f }
 Dir[Rails.root.join("spec/support/**/*.rb")].sort.each { |f| require f }
 
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
