+++ spec/support/search_api_helpers.rb
