+++ spec/support/supergroups_helpers.rb
